Finished.....for now
Play does not collide wth the enemy for some reason. Debug shit needs to
be pulled out. Textures are a litte jacked atm but should be able to be
fixed. Score and lives work but need to be moved to a visible spot on
the sceen. Screen ratio needs to be decided and set. That is the end of
this tutorial. Things that could be implemented within this system:
A new wave system that allows you to quickly prototype wave designs.
Particle effects to go along with the destruction of the enemies.
Power ups and SFX
......